pyproject.toml
[tool.poetry]
name = "llm-kit"
version = "0.1.0"
description = "Generate Data from LLM using LLMKit"
authors = ["Tosin Amuda <tosinamuda@gmail.com>"]
readme = "README.md"
packages = [{include = "llmkit"}]

[tool.poetry.dependencies]
python = "^3.11"
pydantic = "^2.4.2"
openai = "^0.28.1"
python-dotenv = "^1.0.0"


[tool.poetry.group.dev.dependencies]
black = "^23.10.0"
flake8 = "^6.1.0"
pylint = "^3.0.1"

[tool.poetry.scripts]
example = "examples.main"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
